src/security.py
```python
# Importing necessary libraries
import boto3
from azure.identity import DefaultAzureCredential
from azure.core.exceptions import AzureError
from google.cloud import iam
from google.api_core.exceptions import GoogleAPICallError
from hashicorp_vault import VaultClient
from hashicorp_vault.exceptions import VaultError

class Security:
    def __init__(self, model):
        self.model = model
        self.aws_client = boto3.client('iam')
        self.azure_credential = DefaultAzureCredential()
        self.google_client = iam.Client()
        self.vault_client = VaultClient(url="<VAULT_SERVER_URL>", token="<VAULT_TOKEN>")

    def aws_iam(self, policy_name):
        """
        Function to manage AWS IAM policies
        """
        try:
            response = self.aws_client.create_policy(
                PolicyName=policy_name,
                PolicyDocument='string',
                Description='IAM policy for model security',
                Path='/service-role/'
            )
            return response
        except Exception as e:
            print(f"Error in AWS IAM: {e}")
            return None

    def azure_rbac(self, role_name):
        """
        Function to manage Azure RBAC roles
        """
        try:
            role = self.azure_credential.get_token(role_name)
            return role
        except AzureError as e:
            print(f"Error in Azure RBAC: {e}")
            return None

    def google_iam(self, policy_name):
        """
        Function to manage Google IAM policies
        """
        try:
            policy = self.google_client.get_iam_policy(policy_name)
            return policy
        except GoogleAPICallError as e:
            print(f"Error in Google IAM: {e}")
            return None

    def hashicorp_vault(self, secret_path):
        """
        Function to manage HashiCorp Vault secrets
        """
        try:
            secret = self.vault_client.read_secret(secret_path)
            return secret
        except VaultError as e:
            print(f"Error in HashiCorp Vault: {e}")
            return None
```
