tests/test_security.py
```python
# Importing necessary libraries
import unittest
from unittest.mock import patch, MagicMock
from src.security import Security

class TestSecurity(unittest.TestCase):
    def setUp(self):
        self.model = MagicMock()
        self.security = Security(self.model)

    @patch('boto3.client')
    def test_aws_iam(self, mock_client):
        mock_client.return_value.create_policy.return_value = {'Policy': {}}
        response = self.security.aws_iam('test_policy')
        self.assertIsNotNone(response)
        mock_client.return_value.create_policy.assert_called_once()

    @patch('azure.identity.DefaultAzureCredential')
    def test_azure_rbac(self, mock_credential):
        mock_credential.return_value.get_token.return_value = {'token': 'test_token'}
        response = self.security.azure_rbac('test_role')
        self.assertIsNotNone(response)
        mock_credential.return_value.get_token.assert_called_once()

    @patch('google.cloud.iam.Client')
    def test_google_iam(self, mock_client):
        mock_client.return_value.get_iam_policy.return_value = {'policy': {}}
        response = self.security.google_iam('test_policy')
        self.assertIsNotNone(response)
        mock_client.return_value.get_iam_policy.assert_called_once()

    @patch('hashicorp_vault.VaultClient')
    def test_hashicorp_vault(self, mock_client):
        mock_client.return_value.read_secret.return_value = {'secret': {}}
        response = self.security.hashicorp_vault('test_secret_path')
        self.assertIsNotNone(response)
        mock_client.return_value.read_secret.assert_called_once()

if __name__ == '__main__':
    unittest.main()
```
